Resolve "[Pundit] Przygotować model uprawnień na podstawie pundit"

Closes #6

See merge request maciejbiel/todo-graphql!6

+++ app/graphql/mutations/delete_project_mutation.rb
@@ -14,6 +14,8 @@ module Mutations
 
       project = current_user.projects.find(id)
 
+      authorize project, :delete?
+
       if project.destroy
         {
           success: true,
