+++ app/graphql/mutations/delete_task_mutation.rb
@@ -13,9 +13,13 @@ module Mutations
     def resolve(project_id:, id:)
       authenticate_user!
 
-      task = current_user
-             .projects.find(project_id)
-             .tasks.find(id)
+      project = current_user.projects.find(project_id)
+
+      authorize project, :read?
+
+      task = project.tasks.find(id)
+
+      authorize task, :delete?
 
       if task.destroy
         {
